+++ serialization/src/main/java/org/finos/rune/mapper/processor/pruner/PruningStrategy.java
@@ -1,4 +1,4 @@
-package org.finos.rune.mapper.processor;
+package org.finos.rune.mapper.processor.pruner;
 
 /*-
  * ==============
@@ -22,9 +22,6 @@ package org.finos.rune.mapper.processor;
 
 import com.rosetta.model.lib.RosettaModelObjectBuilder;
 
-public class EmptyAttributePruningStrategy implements PruningStrategy {
-    @Override
-    public void prune(RosettaModelObjectBuilder builder) {
-        builder.prune();
-    }
+public interface PruningStrategy {
+    void prune(RosettaModelObjectBuilder builder);
 }
